RFC: Warn that v2 byte-order must match native order
Summary:

When using a barectf 2 (v2) configuration file in barectf 3 (v3), the
native endianness of the system to be traced can be incorrectly assumed.
When the assumed native endianness is incorrect, it results in the
generated tracer producing data with the an endianness that does not
match the endianness expected in the trace's metadata file. So, when the
trace is opened (e.g. with barectf) the trace will appear to be
corrupted.

To reduce the chance of tracer misconfiguration, print a message to warn
users that the v2 byte-order property must match the native endianness
of the system to be traced.

Context:

Barectf 2 configurations require a user-defined property called
'byte-order'. This property defines the desired byte order of the data
to be produced by the generated tracer. Barectf 3 removed the byte-order
property and does not allow the user to specify the byte order of data
produced by the tracer.

Instead, barectf 3 configurations require a new user-defined property
called 'native-byte-order'. This property defines the native endianness
of the system to be traced by the generated tracer. The
native-byte-order also determines the order of the data produced by the
tracer.

To allow backwards compatibility with barectf 2 configuration files,
when parsing a v2 config file barectf 3 defines the required v3
native-byte-order property by copying the value of byte-order from the
v2 config. These two config properties do not mean the same thing. In
other words, the copy assumes that the produced data's desired byte
order matches the system's native byte order.

Problem example:

This native byte order assumption will be false in the unusual and
unrecommended case where the same v2 config is used with barectf 3 to
generate tracers for systems with different native byte orders. For the
systems where the assumed native-byte-order property does not match the
system's actual native byte order, the tracer will be misconfigured and
will generate data with the incorrect endianness.

Changes:

In barectf 3, a v2 config file's byte-order property needs to match the
target system's native endianness for the generated tracer to produce
valid data. To reduce the chance of a mismatch, print message to
communicate endianness more explicitly and to warn users that the v2
byte-order property must match the native endianness of the system to be
traced.

This is not a fix since users are still able to generate, compile, and
use a tracer that does not match the system's native endianness and
therefore that will not work as expected.

Change-Id: I348460ccef1169ab2f025fac93bc60c442ffbd22
Signed-off-by: Erica Bugden <ebugden@efficios.com>